Added css
Added some bottom menu with css and stuff as well as a timer for when start and stop are pressed. Idk what the css all does, I just try stuff until it works and then if you take it out it breaks then I need it. Keeping the design brutally basic until it works and then we can deal with that stuff later.