new polyline each time and null the old one
Trying it slightly different. Making a new polyline each time start is pressed and nulling the old one when stop is pressed. Again Idk if this will keep them all on the screen at once, need to test.
Also moved everything into the on page load bit because the buttons needed the polyline and they were in different scopes