Add files via upload
Switching to using the html speed reading. I did not know that existed so trying it. I used strava to get my walking pace so I'll know if it's a decent output for jogga or not. If this doesn't work well or is not reliable I'll try and pair it with my way and see if I can get the best of both worlds. I did test strava a little and saw that it doesn't update based on time, I think it goes by distance and then updates frequently by dropping old readings and adding new ones or something giving that effect so maybe try that after