added history page
added history.html with bones of a heading, something for when no runs saved, something for when a run is there and the outline of a big pop up called a modal that will be what is showing the run with the map and more info if poss. It will all need done in js at the time of load so this outline is best I can do now.