Add files via upload
it's not saving on the page closing, I have no clue why, it just isnt. Im changing to save when the app is minimised and see if that does anything. It worked earlier and I just have no clue why it doesnt do anything now. It is trying to load from local storage and not finding the file